Update SourceCounts_Upload.ipynb
Fix paths to the source count tables, hopefully so that it can run on jupyter notebook.

The previous upload, btw, was only to plot a couple of figures better. Now the final plots of SFG and AGN, each alone, are normalised w.r.t. the total respective SFG/AGN population count, not the total population count.